Removed libremap dependecies. They are NOT needed by Wibed
Since commit 021b96d7c5c668fbcb5375c65cee90832bb2854f on OpenWRT/Lede,
when you are building the Image, it fails if the dependecies are not
met.

Currently, if you try to compile a Wibed image without adding LibreMap
feed, it'll fail.

Since it's not completly necessary unless you want the nodes to appear
in the map ( and you have to configure them to do it ), I thought it
would be correct to delete the dependency.

If the final user wants to use libremap, he/she can add the feed and
install the package separatedly.